requirements.txt
gunicorn==20.1.0
psycopg2_binary
Werkzeug==2.3.6
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
channels==3.0.5
charset-normalizer==3.2.0
cloudinary==1.33.0
constantly==15.1.0
cryptography==41.0.3
daphne==3.0.2
Django==4.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-eventstream==4.5.1
django-grip==3.4.0
django-redis==5.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
gripcontrol==4.2.0

hyperlink==21.0.0
idna==3.4
incremental==22.10.0
MarkupSafe==2.1.3
Pillow==10.0.0
pubcontrol==3.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytz==2023.3

requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
whitenoise==6.5.0
zope.interface==6.0


